Backend/app/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-RESTX==1.1.0
Flask-Cors==3.0.10
psycopg2-binary==2.9.9
bcrypt==4.0.1
PyJWT==2.7.0
werkzeug<3
